format

+++ fincore.c
@@ -115,7 +115,7 @@ void fincore(char* path,
     int cached = 0;
     int printed = 0;
 
-    int regions[10];
+    long regions[10];
 
     fflush( stdout );
